requirements.txt
# requirements.txt

streamlit==1.12.2
opencv-python-headless==4.6.0.66
#streamlit
#opencv-python==4.7.0.72
numpy
pandas
feat
keras
matplotlib
Pillow
requests
torchvision
tensorflow>=2.0.0

